refactor(config): derive session types from schema
